chore: migrate jsii-diff tests to jest
